+++ templates/chat/next.config.ts
@@ -1,5 +1,15 @@
 import type { NextConfig } from 'next'
 
 export default {
-	reactStrictMode: true
+	reactStrictMode: true,
+	transpilePackages: [
+		'@rubriclab/auth',
+		'@rubriclab/webhooks',
+		// '@rubriclab/actions',
+		// '@rubriclab/blocks',
+		'@rubriclab/agents',
+		'@rubriclab/events'
+		// '@rubriclab/chains',
+		// '@rubriclab/shapes'
+	]
 } satisfies NextConfig